Browser Bookmarks: Fix heap memory errors
