+++ extensions/browser-bookmarks/package.json
@@ -49,9 +49,8 @@
     }
   ],
   "dependencies": {
-    "@raycast/api": "^1.64.0",
-    "@raycast/utils": "^1.10.0",
-    "fuse.js": "^7.0.0",
+    "@raycast/api": "^1.72.1",
+    "@raycast/utils": "^1.15.0",
     "ini": "^4.1.1",
     "simple-plist": "^1.4.0-0",
     "sql.js": "^1.8.0"
