+++ extensions/browser-bookmarks/src/hooks/useChromeBetaBookmarks.ts
@@ -5,11 +5,12 @@ import useChromiumBookmarks from "./useChromiumBookmarks";
 
 const CHROME_BETA_PATH = `${homedir()}/Library/Application Support/Google/Chrome Beta`;
 
-export default function useChromeBetaBookmarks(enabled: boolean) {
+export default function useChromeBetaBookmarks(enabled: boolean, query: string) {
   return useChromiumBookmarks(enabled, {
     path: CHROME_BETA_PATH,
     browserName: "Chrome Beta",
     browserIcon: "chrome-beta.png",
     browserBundleId: BROWSERS_BUNDLE_ID.chromeBeta,
+    query,
   });
 }
