+++ extensions/browser-bookmarks/src/hooks/useEdgeCanaryBookmarks.ts
@@ -5,11 +5,12 @@ import useChromiumBookmarks from "./useChromiumBookmarks";
 
 const EDGECANARY_BOOKMARKS_PATH = `${homedir()}/Library/Application Support/Microsoft Edge Canary`;
 
-export default function useEdgeBookmarks(enabled: boolean) {
+export default function useEdgeBookmarks(enabled: boolean, query: string) {
   return useChromiumBookmarks(enabled, {
     path: EDGECANARY_BOOKMARKS_PATH,
     browserName: "Edge Canary",
     browserIcon: "edgeCanary.png",
     browserBundleId: BROWSERS_BUNDLE_ID.edgeCanary,
+    query,
   });
 }
