+++ extensions/browser-bookmarks/src/hooks/useEdgeDevBookmarks.ts
@@ -5,11 +5,12 @@ import useChromiumBookmarks from "./useChromiumBookmarks";
 
 const EDGEDEV_BOOKMARKS_PATH = `${homedir()}/Library/Application Support/Microsoft Edge Dev`;
 
-export default function useEdgeBookmarks(enabled: boolean) {
+export default function useEdgeBookmarks(enabled: boolean, query: string) {
   return useChromiumBookmarks(enabled, {
     path: EDGEDEV_BOOKMARKS_PATH,
     browserName: "Edge Dev",
     browserIcon: "edgeDev.png",
     browserBundleId: BROWSERS_BUNDLE_ID.edgeDev,
+    query,
   });
 }
