+++ extensions/browser-bookmarks/src/hooks/useVivaldiBrowser.ts
@@ -5,11 +5,12 @@ import useChromiumBookmarks from "./useChromiumBookmarks";
 
 const VIVALDI_BOOKMARKS_PATH = `${homedir()}/Library/Application Support/Vivaldi`;
 
-export default function useVivaldiBookmarks(enabled: boolean) {
+export default function useVivaldiBookmarks(enabled: boolean, query: string) {
   return useChromiumBookmarks(enabled, {
     path: VIVALDI_BOOKMARKS_PATH,
     browserName: "Vivaldi",
     browserIcon: "vivaldi.png",
     browserBundleId: BROWSERS_BUNDLE_ID.vivaldi,
+    query,
   });
 }
